updating dm_env wrapper and training script
